+++ lib/asset/assetmultipart.js
@@ -15,7 +15,7 @@
 require("core-js/stable");
 
 const { IllegalArgumentError } = require("../error");
-const { isPositiveNumber, isValidWebUrl } = require("../util");
+const { isValidWebUrl, isPositiveNumber } = require("../util");
 
 const PRIVATE = Symbol("PRIVATE");
