+++ lib/file.js
@@ -16,7 +16,8 @@ const filterObject = require('filter-obj');
 const fs = require('fs');
 const path = require('path');
 const { downloadStream, uploadStream } = require('./stream');
-const { createReadStream, createWriteStream, getFileStats, isPositiveNumber } = require('./util');
+const { getFileStats, isPositiveNumber } = require('./util');
+const { createReadStream, createWriteStream } = require('./stream-util');
 const { retry } = require('./retry');
 const { HttpStreamError, IllegalArgumentError } = require('./error');
 const { BlockDownload } = require('./block/blockdownload');
