+++ lib/functions/transfer.js
@@ -14,7 +14,8 @@
 
 const logger = require("../logger");
 const { AsyncGeneratorFunction } = require("../generator/function");
-const { isFileProtocol, streamToPooledBuffer } = require("../util");
+const { isFileProtocol } = require("../util");
+const { streamToPooledBuffer } = require("../stream-util");
 const { issuePut, streamGet } = require("../fetch");
 const { retry } = require("../retry");
 const { HTTP } = require("../constants");
