+++ lib/stream.js
@@ -12,10 +12,10 @@
 
 'use strict';
 
-const fetch = require('./fetch');
+const fetch = require("./fetch");
 const { retry } = require("./retry");
-const { parseResourceHeaders } = require('./headers');
-const { HttpStreamError } = require('./error');
+const { parseResourceHeaders } = require("./headers");
+const { HttpStreamError } = require("./error");
 const { HTTP, MIMETYPE } = require("./constants");
 
 /**
