+++ test/util.test.js
@@ -16,64 +16,9 @@
 
 const assert = require("assert");
 const util = require("../lib/util");
-const fs = require('fs').promises;
-const path = require('path');
-const { EventEmitter } = require("events");
-const { TransferMemoryBuffer } = require('../lib/transfer-memory-allocator');
+const path = require("path");
 
 describe("util", function() {
-    it('createReadStream-error', async function() {
-        try {
-            await util.createReadStream("badfile");
-            assert.fail("failure expected");
-        } catch (e) {
-            assert.ok(e.message.includes("ENOENT: no such file or directory"), e.message);
-        }
-    });
-
-    it('creates a read stream', async function() {
-        await fs.writeFile(path.resolve('./test-transfer-file-read-1.dat'), 'hello world 123', 'utf8');
-        const readStream = await util.createReadStream(path.resolve('./test-transfer-file-read-1.dat'));
-
-        assert.ok(readStream.flags === 'r');
-
-        readStream.destroy();
-        assert.ok(readStream.destroyed);
-
-        try {
-            await fs.unlink(path.resolve('./test-transfer-file-read-1.dat'));
-        } catch(e){
-            // ignore clean-up error
-            console.log(e);
-        }
-    });
-
-    it('createWriteStream-error', async function() {
-        try {
-            await util.createWriteStream("badfolder/badfile");
-            assert.fail("failure expected");
-        } catch (e) {
-            assert.ok(e.message.includes("ENOENT: no such file or directory"), e.message);
-        }
-    });
-
-    it('creates a write stream', async function() {
-        //await fs.writeFile(path.resolve('./test-transfer-file.dat'), 'hello world 123', 'utf8');
-        const writeStream = await util.createWriteStream(path.resolve('./test-transfer-file-write-1.dat'));
-
-        assert.ok(writeStream.flags === 'w');
-
-        writeStream.destroy();
-        assert.ok(writeStream.destroyed);
-
-        try {
-            await fs.unlink(path.resolve('./test-transfer-file-write-1.dat'));
-        } catch(e){
-            // ignore clean-up error
-            console.log(e);
-        }
-    });
-
     it('file-protocol-string', function() {
         assert.ok(util.isFileProtocol('file:///path/to/file'));
     });
@@ -92,99 +37,6 @@ describe("util", function() {
 
     it('file-protocol-url-http', function() {
         assert.ok(!util.isFileProtocol(new URL('http://www.host.com')));
-    });
-
-    function configureStream(toExecute) {
-        const stream = new EventEmitter();
-        const origOn = stream.on;
-        stream.on = (name, data) => {
-            origOn.call(stream, name, data);
-            if (name === 'data') {
-                toExecute(stream);
-            }
-        };
-        return stream;
-    }
-
-    it('stream-to-buffer', async function() {
-        const stream = configureStream((stream) => {
-            stream.emit('data', Buffer.from('Hello W'));
-            stream.emit('data', Buffer.from('orld!'));
-            stream.emit('end');
-        });
-
-        const buffer = await util.streamToBuffer("get", "url", 200, stream, 12);
-        assert.deepStrictEqual(buffer.toString(), 'Hello World!');
-    });
-
-    it('stream-to-error-buffer', function() {
-        const stream = configureStream((stream) => {
-            stream.emit('error', 'there was an error!');
-        });
-        assert.rejects(util.streamToBuffer("get", "url", 200, stream, 12));
-    });
-    
-    it('stream-to-unexpectedlength-buffer', function() {
-        const stream = configureStream((stream) => {
-            stream.emit('data', 'test');
-            stream.emit('end');
-        });
-        assert.rejects(util.streamToBuffer("get", "url", 200, stream, 12));
-    });
-
-    it('stream-to-pooled-buffer', async function() {
-        const stream = configureStream((stream) => {
-            stream.emit('data', Buffer.from('Hello W'));
-            stream.emit('data', Buffer.from('orld!'));
-            stream.emit('end');
-        });
-
-        const memoryAllocator = new TransferMemoryBuffer(20);
-        const memoryBufferBlock = await util.streamToPooledBuffer("get", "url", 200, stream, 12, memoryAllocator);
-        assert.deepStrictEqual(memoryBufferBlock.buffer.toString(), 'Hello World!');
-    });
-
-    it('stream-to-error-pooled-buffer', function() {
-        const stream = configureStream((stream) => {
-            stream.emit('error', 'there was an error!');
-        });
-        const memoryAllocator = new TransferMemoryBuffer(15);
-        assert.rejects(util.streamToPooledBuffer("get", "url", 200, stream, 12, memoryAllocator));
-    });
-    
-    it('stream-to-unexpectedlength-pooled-buffer', function() {
-        const stream = configureStream((stream) => {
-            stream.emit('data', 'test');
-            stream.emit('end');
-        });
-        const memoryAllocator = new TransferMemoryBuffer(14);
-        assert.rejects(util.streamToPooledBuffer("get", "url", 200, stream, 12, memoryAllocator));
-    });
-
-    it('stream-to-buffer (stream to pooled buffer fallback when no memory allocator)', async function() {
-        const stream = configureStream((stream) => {
-            stream.emit('data', Buffer.from('Hello W'));
-            stream.emit('data', Buffer.from('orld!'));
-            stream.emit('end');
-        });
-
-        const buffer = await util.streamToPooledBuffer("get", "url", 200, stream, 12);
-        assert.deepEqual(buffer.toString(), 'Hello World!');
-    });
-
-    it('stream-to-error-buffer (stream to pooled buffer fallback when no memory allocator)', function() {
-        const stream = configureStream((stream) => {
-            stream.emit('error', 'there was an error!');
-        });
-        assert.rejects(util.streamToPooledBuffer("get", "url", 200, stream, 12));
-    });
-    
-    it('stream-to-unexpectedlength-buffer (stream to pooled buffer fallback when no memory allocator)', function() {
-        const stream = configureStream((stream) => {
-            stream.emit('data', 'test');
-            stream.emit('end');
-        });
-        assert.rejects(util.streamToPooledBuffer("get", "url", 200, stream, 12));
     });
 
     it('url to path', function() {
